requirements.txt
# Face Recognition and Computer Vision
face_recognition==1.3.0
dlib==19.24.2
opencv-python==4.8.1.78
Pillow==10.0.1
numpy==1.24.3

# Raspberry Pi Hardware Interface
RPi.GPIO==0.7.1
picamera2==0.3.12

# Telegram Bot API
python-telegram-bot==20.6

# Utilities
requests==2.31.0
python-dateutil==2.8.2
schedule==1.2.0

# Logging and Configuration
pyyaml==6.0.1
colorlog==6.7.0

# Web interface (for development/testing)
Flask==2.3.3
Flask-CORS==4.0.0
gunicorn==21.2.0

# Optional: Performance monitoring
psutil==5.9.5
